% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_study_interval_cols.R
\name{add_study_interval_cols}
\alias{add_study_interval_cols}
\title{Adds columns for the study setup}
\usage{
add_study_interval_cols(
  pheno_data,
  exp_age = 30,
  exp_len = 10,
  wash_len = 2,
  out_len = 8
)
}
\arguments{
\item{pheno_data}{A data.frame with at least the columns:
\code{START_OF_FOLLOWUP}, \code{END_OF_FOLLOWUP}, and
\code{DATE_OF_BIRTH}.}

\item{exp_age}{An integer. Age at which exposure period starts
(in years).}

\item{exp_len}{An integer. Length of the exposure period
(in years).}

\item{wash_len}{An integer. Length of the washout period
(in years).}

\item{out_len}{An integer. Length of the observation period
(in years).}
}
\value{
The phenotype data.frame with added columns:
\itemize{
\item FOLLOWUP: lubdridate intervals created using columns
\code{START_OF_FOLLOWUP} and \code{END_OF_FOLLOWUP}.
\item ENDPT_FREE: lubdridate intervals
which ranges from \code{DATE_OF_BIRTH} to the
end of the washout period. In this time
period there should be no endpoint
occurance for eligible study participants.
\item STUDY_TIME: lubdridate intervals which range from
the start of the exposure period to the
end of the observation period.
}
}
\description{
Adds the three columns, \code{FOLLOWUP}, \code{ENDPT_FREE}, and \code{STUDY_TIME}
to the phenotype data.frame. Each column contains lubridate
\code{\link[lubridate]{interval}}s.
}
\details{
For the input data format see:
\href{https://docs.google.com/document/d/1GbZszpPeyf-hyb0V_YDx828YbM7woh8OBJhvzkEwo2g/edit}{INTERVENE Phenotype File Definition}
}
\author{
Kira E. Detrois
}
