% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_age_at_diag}
\alias{calc_age_at_diag}
\title{A helper function to calculate the age at diagnosis.}
\usage{
calc_age_at_diag(bds, diag_dates)
}
\arguments{
\item{bds}{A Date. The birth days of the individuals.}

\item{diag_dates}{A Date. Date of diagnosis.}
}
\description{
A helper function to calculate the age at diagnosis.
}
