% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_endpt_free_time.R
\name{calc_endpt_free_time}
\alias{calc_endpt_free_time}
\title{Calcualtes endpoint free intervals}
\usage{
calc_endpt_free_time(bds, exp_age = 30, exp_len = 10, wash_len = 2)
}
\arguments{
\item{bds}{A Date. The birth days of the individuals.}

\item{exp_age}{An integer. Age at which exposure period starts
(in years).}

\item{exp_len}{An integer. Length of the exposure period
(in years).}

\item{wash_len}{An integer. Length of the washout period
(in years).}
}
\value{
A lubdridate interval. The endpoint free intervals.
}
\description{
Calculates the endpoint free interval for each individual. The
endpoint free interval is the period from birth until the
prediction period begins. Thus, the interval also contains the
exposure and washout period.
}
\details{
Creates lubridate \code{\link[lubridate]{interval}}s.
}
\examples{
bds <- c(as.Date("1923/07/01"), as.Date("1823/07/02"), as.Date("2002/04/01"))
calc_endpt_free_time(bds,
                     exp_age=30,
                     exp_len=10,
                     wash_len=2)

}
\author{
Kira E. Detrois
}
