% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_study_time.R
\name{calc_study_time}
\alias{calc_study_time}
\title{Calculates total study time interval}
\usage{
calc_study_time(bds, study)
}
\arguments{
\item{bds}{A Date. The birth days of the individuals.}

\item{study@obs_len}{An integer. Length of the prediction period.}
}
\value{
A lubdridate interval. The total study time interval.
}
\description{
Calculates the total study time for each individual. The study time
is the time from the begining of the exposure period to the end
of the prediction period.
}
\details{
Creates lubridate \code{\link[lubridate]{interval}}s.
}
\examples{
bds <- c(as.Date("1923/07/01"), as.Date("1823/07/02"), as.Date("2002/04/01"))
study <- new("study", exp_age=30, exp_len=10, wash_len=2, obs_len=8)
calc_study_time(bds, study)

}
\author{
Kira E. Detrois
}
