% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_inputs.R
\name{check_cols_exist}
\alias{check_cols_exist}
\title{Checks that the data.frame contains the chosen columns}
\usage{
check_cols_exist(pheno_data, cols, func_name)
}
\arguments{
\item{pheno_data}{A data.frame.}

\item{cols}{A charachter. The columns \code{pheno_data} needs to contain.}

\item{func_name}{A character. The name of the function to print in the error message.}
}
\description{
Checks that the data.frame contains the chosen columns
}
\author{
Kira E. Detrois
}
