% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_df.R
\name{create_test_df}
\alias{create_test_df}
\title{Creates a data.frame for testing purposes.}
\usage{
create_test_df(n_indv = 10)
}
\arguments{
\item{n_indv}{An integer. Number of individuals to simulate.}
}
\value{
a data.frame with the simulated data.
}
\description{
Creates a data.frame for testing purposes.
}
\author{
Kira E. Detrois
}
