% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample_cntrls.R
\name{downsample_cntrls}
\alias{downsample_cntrls}
\title{Downsamples the number of controls in the data}
\usage{
downsample_cntrls(pheno_data, endpt, downsample_fctr = 4)
}
\arguments{
\item{pheno_data}{A data.frame with at least the column defined in
the variable endpoint.}

\item{endpt}{A string. The column name of the current endpoint of
interest.}

\item{downsample_fctr}{A numeric. Defines how many controls there
should be for every case.}
}
\description{
Downsamples the number of controls in the data
}
\author{
Kira E. Detrois
}
