% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_followup_time.R
\name{get_followup_time}
\alias{get_followup_time}
\title{Gets the follow-up period}
\usage{
get_followup_time(pheno_data)
}
\arguments{
\item{pheno_data}{A data.frame with at least the columns:
\code{END_OF_FOLLOWUP} and \code{START_OF_FOLLOWUP}.}
}
\description{
Gets the follow-up period for each individual as a lubridate
\code{\link[lubridate]{interval}}.
}
\details{
Currently, if there is NAs in the \code{END_OF_FOLLOWUP} column replaces
it with the current date. This assumes that the missingness comes
from the fact that the followup has not yet ended and is not dues
some other reason.
}
\author{
Kira E. Detrois
}
