% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_study_elig_indv.R
\name{get_study_elig_indv}
\alias{get_study_elig_indv}
\title{Gets the individuals which are eligble given the study setup}
\usage{
get_study_elig_indv(
  pheno_data,
  exp_age = 30,
  exp_len = 10,
  wash_len = 2,
  out_len = 8,
  endpt = "J10_ASTHMA",
  downsample_fctr = NA
)
}
\arguments{
\item{pheno_data}{A data.frame with at least the columns:
\code{ID}, \code{SEX}, \code{DATE_OF_BIRTH}, \code{ANCESTRY},
\code{START_OF_FOLLOWUP}, \code{END_OF_FOLLOWUP},
\code{DATE_OF_BIRTH}, and i.e. \code{J10_ASTHMA}, and
\code{J10_ASTHMA_DATE} where the columns are the study
endpoint and date, which will differ depending on
the input variable \code{endpt}.}

\item{exp_age}{An integer. Age at which exposure period starts
(in years).}

\item{exp_len}{An integer. Length of the exposure period
(in years).}

\item{wash_len}{An integer. Length of the washout period
(in years).}

\item{out_len}{An integer. Length of the observation period
(in years).}

\item{endpt}{A string. The column name of the current endpoint of
interest.}

\item{downsample_fctr}{A numeric. Defines how many controls there
should be for every case.
Default is NA, which means no
downsampling is performed.}
}
\value{
A tibble with the information for the eligible individuals
with columns: \code{ID}, \code{SEX}, \code{DATE_OF_BIRTH},
\code{START_OF_FOLLOWUP}, \code{END_OF_FOLLOWUP}, \code{ANCESTRY},
and i.e. \code{J10_ASTHMA}, and \verb{J10_ASTHMA_DATE. where the last two columns are the study endpoint and date,  which will differ depending on the input variable }endpt`.
}
\description{
\itemize{
\item The study setup consist of an exposure window,
a washout period, and a observation period. See function
\code{\link{calc_study_time}}.
\item Eligible individuals must have a follow-up period that
covers the whole study time interval. See function
\code{\link{filter_too_short_followup}}.
\item Eligible individuals cannot have missing date in the
column for the endpoint of interest. See function
\code{\link{filter_missing_endpt_data}}.
\item Eligible individuals cannot have the
selected endpoint of interest inside the endpoint free
period. The endpoint free interval is the period from birth
until the observation period begins. See function
\code{\link{filter_early_endpt}}.

}
}
\author{
Kira E. Detrois
}
