% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_output.R
\name{get_study_file_name}
\alias{get_study_file_name}
\title{Creates a file name for the current study setup}
\usage{
get_study_file_name(envir)
}
\arguments{
\item{envir}{A list with at least entries \code{endpt}, \code{exp_age}, \code{exp_age},
\code{exp_age}, and \code{exp_age}.}
}
\description{
Creates a file name for the current study setup
}
\author{
Kira E. Detrois
}
