% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_output.R
\name{log_msg_string}
\alias{log_msg_string}
\title{Creates a string of the current study setup}
\usage{
log_msg_string(elig_indv, study)
}
\arguments{
\item{elig_indv}{A tibble with the information for the eligible
individuals that should be written to the file.}

\item{study}{An S4 class representing the study setup.}
}
\value{
The log message character string
}
\description{
Creates a string of the current study setup
}
\author{
Kira E. Detrois
}
