% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print_log_msg}
\alias{print_log_msg}
\title{Creates a string of the current study setup}
\usage{
print_log_msg(envir)
}
\arguments{
\item{envir}{A list. The environment of the \code{create_test_df} function.}
}
\description{
Only used inside \code{write_log} function.
}
