% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_class_def.R
\docType{class}
\name{study-class}
\alias{study-class}
\alias{study}
\title{An S4 class representing the study setup}
\description{
An S4 class representing the study setup
}
\section{Slots}{

\describe{
\item{\code{data}}{The actual data.frame with eligible individuals under the
current study setup.}

\item{\code{exp_age}}{An integer. Age at which exposure period starts (in years)}

\item{\code{exp_len}}{asdf}

\item{\code{wash_len}}{as}

\item{\code{n_cases}}{asdf}

\item{\code{n_cntrls}}{asdf}
}}

