% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_inputs.R
\name{test_endpt_correct}
\alias{test_endpt_correct}
\title{Tests that the input for endpt is of the correct type.
Meaning, a string and not a vector of strings. Also tests
that the endpoint is actually included in the data.frame}
\usage{
test_endpt_correct(endpt, cols)
}
\arguments{
\item{endpt}{A character. The endpoint of interest.}

\item{cols}{A character. The columns of the data.frame.}
}
\description{
Tests that the input for endpt is of the correct type.
Meaning, a string and not a vector of strings. Also tests
that the endpoint is actually included in the data.frame
}
