% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_inputs.R
\name{test_input_vars_correct}
\alias{test_input_vars_correct}
\title{Tests that the inputs into the function
\code{\link{get_study_elig_indv}} are of correct type and that
the endpoint selected is a column in the data.frame.}
\usage{
test_input_vars_correct(envir)
}
\arguments{
\item{envir}{A list.}
}
\description{
Tests that the inputs into the function
\code{\link{get_study_elig_indv}} are of correct type and that
the endpoint selected is a column in the data.frame.
}
