% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_inputs.R
\name{test_length_vars_are_integers}
\alias{test_length_vars_are_integers}
\title{Tests that the variable is an integer.}
\usage{
test_length_vars_are_integers(var)
}
\arguments{
\item{var}{A numeric or integer.}
}
\description{
Tests that the variable is an integer.
}
