% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_output.R
\name{write_log_file}
\alias{write_log_file}
\title{Prints the current study setup or writes it to a file}
\usage{
write_log_file(envir)
}
\arguments{
\item{envir}{A list with at least entries \code{pheno_data}, \code{endpt},
\code{exp_age}, \code{exp_len}, \code{wash_len}, and \code{obs_len}.}
}
\description{
Prints the current study setup or writes it to a file
}
