% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_log_file_path}
\alias{write_log_file_path}
\title{Prints the current study setup or writes it to a file}
\usage{
write_log_file_path(envir)
}
\arguments{
\item{envir}{A list. The environment of the \code{create_test_df} function.}
}
\description{
Only used inside \code{create_test_df} function.
}
