% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_output.R
\name{write_res_file}
\alias{write_res_file}
\title{Writes results to a tab-delim file}
\usage{
write_res_file(study, write_res, res_dir)
}
\arguments{
\item{study}{A study object with the current study setup.}

\item{write_res}{A boolean. Defines whether to write the results to
a file or not. Default is FALSE.}

\item{res_dir}{A character. The directory to write the results to.}
}
\description{
Writes results to a tab-delim file
}
\author{
Kira E. Detrois
}
