% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_to_file.R
\name{write_to_file}
\alias{write_to_file}
\title{Writes results to a tab-delim file}
\usage{
write_to_file(
  final_data,
  file_dir = paste0(getwd(), "/results/"),
  file_name = "eligible_indv"
)
}
\arguments{
\item{final_data}{The data.frame to be saved.}

\item{file_dir}{A character. Directory path where file should be
written to.}

\item{file_name}{A character. File name, without file-type ending.
so do not add i.e. ".tsv" to the name.}
}
\description{
Writes results to a tab-delim file
}
\author{
Kira E. Detrois
}
