% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_study_interval_cols.R
\name{add_study_interval_cols}
\alias{add_study_interval_cols}
\title{Adds columns for the study setup}
\usage{
add_study_interval_cols(pheno_data, study)
}
\arguments{
\item{pheno_data}{A data.frame with at least the column \code{DATE_OF_BIRTH}.}

\item{study}{A study object with the current study setup.}
}
\value{
The phenotype data.frame with added columns:
\itemize{
\item \code{ENDPT_FREE}: lubdridate intervals
which ranges from \code{DATE_OF_BIRTH} to the
end of the washout period. In this time
period there should be no endpoint
occurance for eligible study participants.
\item \code{STUDY_TIME}: lubdridate intervals which range from
the start of the exposure period to the
end of the prediction period.
}
}
\description{
Adds the three columns, \code{FOLLOWUP}, \code{ENDPT_FREE}, and \code{STUDY_TIME}
to the phenotype data.frame. Each column contains lubridate
\code{\link[lubridate]{interval}}s.
}
\details{
For the input data format see:
\href{https://docs.google.com/document/d/1GbZszpPeyf-hyb0V_YDx828YbM7woh8OBJhvzkEwo2g/edit}{INTERVENE Phenotype File Definition}
}
\author{
Kira E. Detrois
}
