% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_end_of_study.R
\name{calc_end_of_study}
\alias{calc_end_of_study}
\title{Calculates the end of the study time}
\usage{
calc_end_of_study(bds, study)
}
\arguments{
\item{bds}{A Date. The birth days of the individuals.}
}
\value{
A lubdridate interval. The total study time interval.
}
\description{
Calculates the end of the study time for each individual. The study time
is the time from the begining of the exposure period to the end
of the prediction period.
}
\examples{
bds <- c(as.Date("1923/07/01"), as.Date("1823/07/02"), as.Date("2002/04/01"))
study <- new("study", exp_age=30, exp_len=10, wash_len=2, obs_len=8)
calc_end_of_study(bds, study)

}
\author{
Kira E. Detrois
}
