% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_onset_time.R
\name{calc_onset_time}
\alias{calc_onset_time}
\title{Calculates the date of onset and the age in days}
\usage{
calc_onset_time(pheno_data, endpt, exp_age, exp_len, wash_len, obs_len)
}
\arguments{
\item{pheno_data}{A data.frame with at least the columns:
\code{ID}, \code{SEX}, \code{DATE_OF_BIRTH}, \code{ANCESTRY},
\code{START_OF_FOLLOWUP}, \code{END_OF_FOLLOWUP},
\code{DATE_OF_BIRTH}, and i.e. \code{J10_ASTHMA}, and
\code{J10_ASTHMA_DATE} where the columns are the study
endpoint and date, which will differ depending on
the input variable \code{endpt}.}

\item{endpt}{A string. The column name of the current endpoint of
interest.}

\item{exp_age}{An integer. Age at which exposure period starts
(in years).}

\item{exp_len}{An integer. Length of the exposure period
(in years).}

\item{wash_len}{An integer. Length of the washout period
(in years).}

\item{obs_len}{An integer. Length of the prediction period
(in years).}
}
\value{
A list(\code{age_days}, \code{age_date}).
\itemize{
\item \code{age_days}: The age at onset in days
\item \code{age_date}: The date of onset
}
}
\description{
Calculates the age at onset of the endpoint of interest.
The date of onset is either the date of diagnosis for cases or the
end of the study period for controls.
}
\examples{
bds <- c(as.Date("1923/07/01"), as.Date("1823/07/02"), as.Date("2002/04/01"))
calc_study_time(bds,
                exp_age=30,
                exp_len=10,
                wash_len=2,
                obs_len=8)

}
\author{
Kira E. Detrois
}
