% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_output.R
\name{get_study_file_name}
\alias{get_study_file_name}
\title{Creates a file name for the current study setup}
\usage{
get_study_file_name(study)
}
\arguments{
\item{study}{A study object with the current study setup.}
}
\description{
Creates a file name for the current study setup
}
\author{
Kira E. Detrois
}
