% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_output.R
\name{log_msg_string}
\alias{log_msg_string}
\title{Creates a string of the current study setup}
\usage{
log_msg_string(envir)
}
\arguments{
\item{envir}{A list with at least entries \code{pheno_data}, \code{endpt},
\code{exp_age}, \code{exp_len}, \code{wash_len}, and \code{obs_len}.}
}
\description{
Only used inside \code{write_res} function.
}
