% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log_msg_string}
\alias{log_msg_string}
\title{Creates a string of the current study setup}
\usage{
log_msg_string(envir)
}
\arguments{
\item{envir}{A list. The environment of the \code{create_test_df} function.}
}
\description{
Only used inside \code{write_log} function.
}
