% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_output.R
\name{print_log_msg}
\alias{print_log_msg}
\title{Creates a string of the current study setup}
\usage{
print_log_msg(envir)
}
\arguments{
\item{envir}{A list with at least entries \code{pheno_data}, \code{endpt},
\code{exp_age}, \code{exp_len}, \code{wash_len}, and \code{obs_len}.}
}
\description{
Only used inside \code{write_log} function.
}
