% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_inputs.R
\name{test_length_vars_are_integers}
\alias{test_length_vars_are_integers}
\title{Tests that the length variables of the function call are integers.}
\usage{
test_length_vars_are_integers(envir)
}
\arguments{
\item{envir}{A list. The function call variables.}
}
\description{
Tests that the length variables of the function call are integers.
}
