% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_output.R
\name{write_res_file}
\alias{write_res_file}
\title{Prints the current study setup or writes it to a file}
\usage{
write_res_file(envir)
}
\arguments{
\item{envir}{A list with at least entries \code{pheno_data}, \code{endpt},
\code{exp_age}, \code{exp_age}, \code{exp_age}, and \code{exp_age}.}
}
\description{
Prints the current study setup or writes it to a file
}
